get_sigcell_simple

a function to determine statistically trait-enriched cell by permutation test.

get_sigcell_simple(
  knn_sparse_mat,
  seed_idx,
  topseed_npscore,
  permutation_times = 1000,
  true_cell_significance = 0.05,
  rda_output = FALSE,
  out_rda = tempfile(fileext = "true_cell_df.rda"),
  mycores = 1,
  rw_gamma = 0.05
)

Arguments

knn_sparse_mat

a sparse matrix used for network propagation, which indicates the adjacent matrix (m x m, where m is the cell number) of cell-to-cell network (M-kNN graph)

seed_idx

a logical vector indicating seed cells (TRUE) and non-seed cells (FALSE) with length of m, where m is the cell number. The length and position are corresponding to knn_sparse_mat

topseed_npscore

a numeric vector of real network propagation score

permutation_times

an integer describe times of permutation test for each cell

true_cell_significance

a numeric value between 0-1 indicating the significant threshold used to determine statistically trait-enriched cell

rda_output

if output details of each permutation as a rda

out_rda

if rda_output=T, an rda will be outputed with path and name specified

mycores

how many cores used for permutation test

rw_gamma

gamma from randomWalk_sparse function. Keep the same with what you used in the real setting

Value

A dataframe with two columns, the first one is seed index (the same with input), the second one is significance from permutation test. In addition, an R data is generated, which contains this dataframe and another dataframe depicting the network propagation score for each cell at each permuation.

Examples

knn_sparse_mat <- example_results$mutualknn30
mono_mat <- example_results$mono_mat
ture_cell_df <- get_sigcell_simple(knn_sparse_mat=knn_sparse_mat,
                                   seed_idx=mono_mat$seed_idx,
                                   topseed_npscore=mono_mat$np_score,
                                   permutation_times=100)
#> Get started!
#> Stationary step: 91
#> Stationary Delta: 9.57457425881841e-06
#> Stationary step: 92
#> Stationary Delta: 9.54680230587292e-06
#> Stationary step: 90
#> Stationary Delta: 9.76276077321982e-06
#> Stationary step: 88
#> Stationary Delta: 9.93007104556721e-06
#> Stationary step: 90
#> Stationary Delta: 9.62229311567847e-06
#> Stationary step: 90
#> Stationary Delta: 9.47772615096603e-06
#> Stationary step: 89
#> Stationary Delta: 9.43320631817339e-06
#> Stationary step: 88
#> Stationary Delta: 9.43772163829556e-06
#> Stationary step: 93
#> Stationary Delta: 9.61059766787955e-06
#> Stationary step: 89
#> Stationary Delta: 9.96263030773154e-06
#> Stationary step: 87
#> Stationary Delta: 9.63625972075843e-06
#> Stationary step: 88
#> Stationary Delta: 9.59920752135556e-06
#> Stationary step: 91
#> Stationary Delta: 9.80091604162607e-06
#> Stationary step: 90
#> Stationary Delta: 9.99026687547467e-06
#> Stationary step: 90
#> Stationary Delta: 9.75525486086529e-06
#> Stationary step: 88
#> Stationary Delta: 9.94881484212702e-06
#> Stationary step: 88
#> Stationary Delta: 9.82321066005993e-06
#> Stationary step: 93
#> Stationary Delta: 9.9440301862905e-06
#> Stationary step: 86
#> Stationary Delta: 9.36423619528793e-06
#> Stationary step: 88
#> Stationary Delta: 9.73072136470244e-06
#> Stationary step: 88
#> Stationary Delta: 9.50374923469248e-06
#> Stationary step: 88
#> Stationary Delta: 9.3671671378618e-06
#> Stationary step: 87
#> Stationary Delta: 9.57618993577492e-06
#> Stationary step: 90
#> Stationary Delta: 9.70781933806319e-06
#> Stationary step: 89
#> Stationary Delta: 9.91365394745394e-06
#> Stationary step: 85
#> Stationary Delta: 9.90869585368558e-06
#> Stationary step: 87
#> Stationary Delta: 9.8535207414702e-06
#> Stationary step: 88
#> Stationary Delta: 9.87273630995817e-06
#> Stationary step: 93
#> Stationary Delta: 9.4760021830699e-06
#> Stationary step: 86
#> Stationary Delta: 9.64163403635478e-06
#> Stationary step: 90
#> Stationary Delta: 9.30699776378942e-06
#> Stationary step: 88
#> Stationary Delta: 9.99959909653895e-06
#> Stationary step: 90
#> Stationary Delta: 9.44890298765176e-06
#> Stationary step: 85
#> Stationary Delta: 9.45328912910634e-06
#> Stationary step: 91
#> Stationary Delta: 9.34448718056681e-06
#> Stationary step: 87
#> Stationary Delta: 9.93457462139834e-06
#> Stationary step: 88
#> Stationary Delta: 9.93898364211985e-06
#> Stationary step: 84
#> Stationary Delta: 9.93782091732426e-06
#> Stationary step: 89
#> Stationary Delta: 9.95551340182591e-06
#> Stationary step: 91
#> Stationary Delta: 9.72550214798095e-06
#> Stationary step: 93
#> Stationary Delta: 9.7175853113511e-06
#> Stationary step: 89
#> Stationary Delta: 9.37257645393851e-06
#> Stationary step: 87
#> Stationary Delta: 9.33835239215376e-06
#> Stationary step: 90
#> Stationary Delta: 9.38026913130519e-06
#> Stationary step: 86
#> Stationary Delta: 9.90610561860761e-06
#> Stationary step: 89
#> Stationary Delta: 9.74781130563429e-06
#> Stationary step: 89
#> Stationary Delta: 9.45185684972928e-06
#> Stationary step: 89
#> Stationary Delta: 9.50076175772539e-06
#> Stationary step: 89
#> Stationary Delta: 9.94802804464137e-06
#> Stationary step: 85
#> Stationary Delta: 9.81076172205276e-06
#> Stationary step: 89
#> Stationary Delta: 9.41253558147759e-06
#> Stationary step: 88
#> Stationary Delta: 9.46561879498362e-06
#> Stationary step: 90
#> Stationary Delta: 9.66292704870433e-06
#> Stationary step: 88
#> Stationary Delta: 9.81642101062862e-06
#> Stationary step: 88
#> Stationary Delta: 9.52898111857366e-06
#> Stationary step: 89
#> Stationary Delta: 9.81709428140763e-06
#> Stationary step: 87
#> Stationary Delta: 9.28593303640789e-06
#> Stationary step: 89
#> Stationary Delta: 9.39973164507114e-06
#> Stationary step: 91
#> Stationary Delta: 9.65091458633059e-06
#> Stationary step: 83
#> Stationary Delta: 9.82715889843375e-06
#> Stationary step: 87
#> Stationary Delta: 9.49710846273424e-06
#> Stationary step: 88
#> Stationary Delta: 9.81066718243302e-06
#> Stationary step: 89
#> Stationary Delta: 9.36618238509533e-06
#> Stationary step: 88
#> Stationary Delta: 9.60203031908236e-06
#> Stationary step: 91
#> Stationary Delta: 9.30803097967021e-06
#> Stationary step: 88
#> Stationary Delta: 9.29872060683426e-06
#> Stationary step: 93
#> Stationary Delta: 9.85990286673576e-06
#> Stationary step: 86
#> Stationary Delta: 9.68571076070491e-06
#> Stationary step: 90
#> Stationary Delta: 9.59537818109605e-06
#> Stationary step: 96
#> Stationary Delta: 9.52551607487604e-06
#> Stationary step: 89
#> Stationary Delta: 9.76064246944989e-06
#> Stationary step: 90
#> Stationary Delta: 9.6102171312179e-06
#> Stationary step: 87
#> Stationary Delta: 9.37119896583585e-06
#> Stationary step: 86
#> Stationary Delta: 9.45301661997356e-06
#> Stationary step: 92
#> Stationary Delta: 9.29935369876151e-06
#> Stationary step: 83
#> Stationary Delta: 9.41995157609904e-06
#> Stationary step: 90
#> Stationary Delta: 9.9421628287053e-06
#> Stationary step: 90
#> Stationary Delta: 9.86518729108976e-06
#> Stationary step: 84
#> Stationary Delta: 9.5270561355081e-06
#> Stationary step: 92
#> Stationary Delta: 9.37290170253184e-06
#> Stationary step: 87
#> Stationary Delta: 9.62333769887339e-06
#> Stationary step: 85
#> Stationary Delta: 9.84201782803479e-06
#> Stationary step: 85
#> Stationary Delta: 9.95286564439916e-06
#> Stationary step: 85
#> Stationary Delta: 9.43479087507097e-06
#> Stationary step: 86
#> Stationary Delta: 9.93909223694752e-06
#> Stationary step: 89
#> Stationary Delta: 9.57607863990439e-06
#> Stationary step: 89
#> Stationary Delta: 9.35095785481315e-06
#> Stationary step: 87
#> Stationary Delta: 9.84486559345049e-06
#> Stationary step: 90
#> Stationary Delta: 9.66336023680439e-06
#> Stationary step: 92
#> Stationary Delta: 9.32949348741077e-06
#> Stationary step: 88
#> Stationary Delta: 9.48485679976563e-06
#> Stationary step: 89
#> Stationary Delta: 9.56120330914006e-06
#> Stationary step: 85
#> Stationary Delta: 9.76013210641229e-06
#> Stationary step: 87
#> Stationary Delta: 9.58612548931663e-06
#> Stationary step: 86
#> Stationary Delta: 9.79474863262497e-06
#> Stationary step: 85
#> Stationary Delta: 9.53180774107278e-06
#> Stationary step: 90
#> Stationary Delta: 9.45981082792115e-06
#> Stationary step: 92
#> Stationary Delta: 9.73933680113664e-06
#> Stationary step: 83
#> Stationary Delta: 9.82011538865364e-06
#> Stationary step: 88
#> Stationary Delta: 9.65385328115969e-06
#> 100
#> cells passed 0.001 threshold: 13.13%
#> cells passed 0.01 threshold: 16.37%
#> cells passed 0.05 threshold: 25.6%
#> your emprical P value threshold: 0.05
#> what propertion of enriched cells over all cells: 25.6%
#> what propertion of seed cells that are enriched cells: 97.81%
#> fold of true cell over seed: 5.1