get_sigcell_simple

a function to determine statistically trait-enriched cell by permutation test.

get_sigcell_simple(
  knn_sparse_mat,
  seed_idx,
  topseed_npscore,
  permutation_times = 1000,
  true_cell_significance = 0.05,
  rda_output = FALSE,
  out_rda = tempfile(fileext = "true_cell_df.rda"),
  mycores = 1,
  rw_gamma = 0.05
)

Arguments

knn_sparse_mat

a sparse matrix used for network propagation, which indicates the adjacent matrix (m x m, where m is the cell number) of cell-to-cell network (M-kNN graph)

seed_idx

a logical vector indicating seed cells (TRUE) and non-seed cells (FALSE) with length of m, where m is the cell number. The length and position are corresponding to knn_sparse_mat

topseed_npscore

a numeric vector of real network propagation score

permutation_times

an integer describe times of permutation test for each cell

true_cell_significance

a numeric value between 0-1 indicating the significant threshold used to determine statistically trait-enriched cell

rda_output

if output details of each permutation as a rda

out_rda

if rda_output=T, an rda will be outputed with path and name specified

mycores

how many cores used for permutation test

rw_gamma

gamma from randomWalk_sparse function. Keep the same with what you used in the real setting

Value

A dataframe with two columns, the first one is seed index (the same with input), the second one is significance from permutation test. In addition, an R data is generated, which contains this dataframe and another dataframe depicting the network propagation score for each cell at each permuation.

Examples

knn_sparse_mat <- example_results$mutualknn30
mono_mat <- example_results$mono_mat
ture_cell_df <- get_sigcell_simple(knn_sparse_mat=knn_sparse_mat,
                                   seed_idx=mono_mat$seed_idx,
                                   topseed_npscore=mono_mat$np_score,
                                   permutation_times=100)
#> Get started!
#> Stationary step: 92
#> Stationary Delta: 9.96363004748587e-06
#> Stationary step: 90
#> Stationary Delta: 9.88723358360506e-06
#> Stationary step: 90
#> Stationary Delta: 9.48756114563026e-06
#> Stationary step: 86
#> Stationary Delta: 9.60748995750898e-06
#> Stationary step: 90
#> Stationary Delta: 9.30495216570052e-06
#> Stationary step: 88
#> Stationary Delta: 9.45895818326023e-06
#> Stationary step: 91
#> Stationary Delta: 9.42346799952073e-06
#> Stationary step: 92
#> Stationary Delta: 9.4375867387001e-06
#> Stationary step: 83
#> Stationary Delta: 9.33715158750476e-06
#> Stationary step: 90
#> Stationary Delta: 9.79224590759526e-06
#> Stationary step: 86
#> Stationary Delta: 9.87104143027189e-06
#> Stationary step: 87
#> Stationary Delta: 9.76323397771106e-06
#> Stationary step: 91
#> Stationary Delta: 9.79306835299441e-06
#> Stationary step: 86
#> Stationary Delta: 9.95832334827472e-06
#> Stationary step: 86
#> Stationary Delta: 9.40718666138102e-06
#> Stationary step: 91
#> Stationary Delta: 9.82027750454571e-06
#> Stationary step: 85
#> Stationary Delta: 9.73379206052872e-06
#> Stationary step: 82
#> Stationary Delta: 9.94897599298537e-06
#> Stationary step: 91
#> Stationary Delta: 9.97936328055525e-06
#> Stationary step: 90
#> Stationary Delta: 9.37992140097643e-06
#> Stationary step: 89
#> Stationary Delta: 9.86456260125549e-06
#> Stationary step: 91
#> Stationary Delta: 9.52388690978328e-06
#> Stationary step: 93
#> Stationary Delta: 9.57597599449186e-06
#> Stationary step: 92
#> Stationary Delta: 9.49047703700981e-06
#> Stationary step: 87
#> Stationary Delta: 9.87806701299995e-06
#> Stationary step: 86
#> Stationary Delta: 9.61910743913735e-06
#> Stationary step: 89
#> Stationary Delta: 9.71927906828521e-06
#> Stationary step: 82
#> Stationary Delta: 9.48667923736612e-06
#> Stationary step: 89
#> Stationary Delta: 9.64330346399671e-06
#> Stationary step: 87
#> Stationary Delta: 9.91937004573877e-06
#> Stationary step: 86
#> Stationary Delta: 9.89829957621106e-06
#> Stationary step: 87
#> Stationary Delta: 9.22414458240106e-06
#> Stationary step: 89
#> Stationary Delta: 9.42651635018793e-06
#> Stationary step: 86
#> Stationary Delta: 9.65023294392838e-06
#> Stationary step: 84
#> Stationary Delta: 9.98285423392133e-06
#> Stationary step: 93
#> Stationary Delta: 9.62077957605981e-06
#> Stationary step: 90
#> Stationary Delta: 9.65764515025664e-06
#> Stationary step: 88
#> Stationary Delta: 9.88681677674632e-06
#> Stationary step: 89
#> Stationary Delta: 9.37854436492598e-06
#> Stationary step: 87
#> Stationary Delta: 9.56623133994612e-06
#> Stationary step: 93
#> Stationary Delta: 9.44868521607277e-06
#> Stationary step: 89
#> Stationary Delta: 9.68988859797643e-06
#> Stationary step: 88
#> Stationary Delta: 9.55290531283618e-06
#> Stationary step: 84
#> Stationary Delta: 9.34402586356639e-06
#> Stationary step: 89
#> Stationary Delta: 9.92719838955695e-06
#> Stationary step: 90
#> Stationary Delta: 9.58105559288028e-06
#> Stationary step: 89
#> Stationary Delta: 9.61553055919952e-06
#> Stationary step: 87
#> Stationary Delta: 9.54111906977343e-06
#> Stationary step: 93
#> Stationary Delta: 9.79781286355831e-06
#> Stationary step: 87
#> Stationary Delta: 9.85252326922154e-06
#> Stationary step: 88
#> Stationary Delta: 9.71283679701306e-06
#> Stationary step: 84
#> Stationary Delta: 9.99766169560593e-06
#> Stationary step: 89
#> Stationary Delta: 9.47179822900079e-06
#> Stationary step: 87
#> Stationary Delta: 9.89268258505701e-06
#> Stationary step: 89
#> Stationary Delta: 9.6598750364313e-06
#> Stationary step: 87
#> Stationary Delta: 9.51089401417507e-06
#> Stationary step: 88
#> Stationary Delta: 9.80848450803237e-06
#> Stationary step: 91
#> Stationary Delta: 9.8521191912268e-06
#> Stationary step: 87
#> Stationary Delta: 9.80368023073606e-06
#> Stationary step: 89
#> Stationary Delta: 9.86259468512932e-06
#> Stationary step: 90
#> Stationary Delta: 9.53137911569065e-06
#> Stationary step: 92
#> Stationary Delta: 9.69708892062068e-06
#> Stationary step: 93
#> Stationary Delta: 9.30214125818054e-06
#> Stationary step: 91
#> Stationary Delta: 9.35516243648118e-06
#> Stationary step: 84
#> Stationary Delta: 9.7753727159729e-06
#> Stationary step: 89
#> Stationary Delta: 9.87364287892505e-06
#> Stationary step: 89
#> Stationary Delta: 9.76711483802963e-06
#> Stationary step: 90
#> Stationary Delta: 9.85360518210783e-06
#> Stationary step: 82
#> Stationary Delta: 9.77750625314143e-06
#> Stationary step: 90
#> Stationary Delta: 9.90518827746851e-06
#> Stationary step: 92
#> Stationary Delta: 9.93297847465633e-06
#> Stationary step: 87
#> Stationary Delta: 9.46281127235886e-06
#> Stationary step: 85
#> Stationary Delta: 9.92735375871481e-06
#> Stationary step: 89
#> Stationary Delta: 9.49744234573282e-06
#> Stationary step: 86
#> Stationary Delta: 9.42697241526764e-06
#> Stationary step: 87
#> Stationary Delta: 9.96438595447216e-06
#> Stationary step: 90
#> Stationary Delta: 9.84437425815741e-06
#> Stationary step: 87
#> Stationary Delta: 9.62599374795442e-06
#> Stationary step: 86
#> Stationary Delta: 9.36322312667441e-06
#> Stationary step: 87
#> Stationary Delta: 9.54676320760144e-06
#> Stationary step: 89
#> Stationary Delta: 9.88649293567513e-06
#> Stationary step: 87
#> Stationary Delta: 9.62456368047186e-06
#> Stationary step: 87
#> Stationary Delta: 9.4309725416601e-06
#> Stationary step: 91
#> Stationary Delta: 9.37977828016102e-06
#> Stationary step: 90
#> Stationary Delta: 9.44320246512476e-06
#> Stationary step: 85
#> Stationary Delta: 9.79875575505178e-06
#> Stationary step: 89
#> Stationary Delta: 9.96752939247404e-06
#> Stationary step: 86
#> Stationary Delta: 9.6164976220067e-06
#> Stationary step: 91
#> Stationary Delta: 9.38873680104584e-06
#> Stationary step: 85
#> Stationary Delta: 9.3277898908566e-06
#> Stationary step: 91
#> Stationary Delta: 9.51026375206487e-06
#> Stationary step: 88
#> Stationary Delta: 9.6370405554026e-06
#> Stationary step: 85
#> Stationary Delta: 9.75618641840196e-06
#> Stationary step: 89
#> Stationary Delta: 9.79625992560366e-06
#> Stationary step: 92
#> Stationary Delta: 9.69235453989373e-06
#> Stationary step: 85
#> Stationary Delta: 9.70252137542744e-06
#> Stationary step: 89
#> Stationary Delta: 9.64632408932021e-06
#> Stationary step: 89
#> Stationary Delta: 9.45648363830636e-06
#> Stationary step: 89
#> Stationary Delta: 9.35325423161361e-06
#> Stationary step: 87
#> Stationary Delta: 9.91002948734427e-06
#> 100
#> cells passed 0.001 threshold: 13.06%
#> cells passed 0.01 threshold: 16.88%
#> cells passed 0.05 threshold: 25.8%
#> your emprical P value threshold: 0.05
#> what propertion of enriched cells over all cells: 25.8%
#> what propertion of seed cells that are enriched cells: 97.81%
#> fold of true cell over seed: 5.14