get_sigcell_simple

a function to determine statistically trait-enriched cell by permutation test.

get_sigcell_simple(
  knn_sparse_mat,
  seed_idx,
  topseed_npscore,
  permutation_times = 1000,
  true_cell_significance = 0.05,
  rda_output = FALSE,
  out_rda = tempfile(fileext = "true_cell_df.rda"),
  mycores = 1,
  rw_gamma = 0.05
)

Arguments

knn_sparse_mat

a sparse matrix used for network propagation, which indicates the adjacent matrix (m x m, where m is the cell number) of cell-to-cell network (M-kNN graph)

seed_idx

a logical vector indicating seed cells (TRUE) and non-seed cells (FALSE) with length of m, where m is the cell number. The length and position are corresponding to knn_sparse_mat

topseed_npscore

a numeric vector of real network propagation score

permutation_times

an integer describe times of permutation test for each cell

true_cell_significance

a numeric value between 0-1 indicating the significant threshold used to determine statistically trait-enriched cell

rda_output

if output details of each permutation as a rda

out_rda

if rda_output=T, an rda will be outputed with path and name specified

mycores

how many cores used for permutation test

rw_gamma

gamma from randomWalk_sparse function. Keep the same with what you used in the real setting

Value

A dataframe with two columns, the first one is seed index (the same with input), the second one is significance from permutation test. In addition, an R data is generated, which contains this dataframe and another dataframe depicting the network propagation score for each cell at each permuation.

Examples

knn_sparse_mat <- example_results$mutualknn30
mono_mat <- example_results$mono_mat
ture_cell_df <- get_sigcell_simple(knn_sparse_mat=knn_sparse_mat,
                                   seed_idx=mono_mat$seed_idx,
                                   topseed_npscore=mono_mat$np_score,
                                   permutation_times=100)
#> Get started!
#> Stationary step: 87
#> Stationary Delta: 9.42426803730815e-06
#> Stationary step: 86
#> Stationary Delta: 9.90072745396192e-06
#> Stationary step: 90
#> Stationary Delta: 9.65505959280439e-06
#> Stationary step: 89
#> Stationary Delta: 9.56572427856387e-06
#> Stationary step: 87
#> Stationary Delta: 9.58192792961716e-06
#> Stationary step: 87
#> Stationary Delta: 9.82658647531104e-06
#> Stationary step: 87
#> Stationary Delta: 9.87627085497902e-06
#> Stationary step: 87
#> Stationary Delta: 9.58771902606309e-06
#> Stationary step: 91
#> Stationary Delta: 9.64139424914445e-06
#> Stationary step: 92
#> Stationary Delta: 9.57092028281049e-06
#> Stationary step: 93
#> Stationary Delta: 9.47265775943068e-06
#> Stationary step: 89
#> Stationary Delta: 9.92225535825735e-06
#> Stationary step: 89
#> Stationary Delta: 9.98174100096028e-06
#> Stationary step: 89
#> Stationary Delta: 9.32563101204927e-06
#> Stationary step: 89
#> Stationary Delta: 9.61257170575795e-06
#> Stationary step: 89
#> Stationary Delta: 9.28348594640894e-06
#> Stationary step: 87
#> Stationary Delta: 9.86396664751874e-06
#> Stationary step: 88
#> Stationary Delta: 9.46217363452761e-06
#> Stationary step: 86
#> Stationary Delta: 9.99354853808956e-06
#> Stationary step: 92
#> Stationary Delta: 9.89402355224409e-06
#> Stationary step: 89
#> Stationary Delta: 9.6852529715502e-06
#> Stationary step: 88
#> Stationary Delta: 9.5932363653375e-06
#> Stationary step: 86
#> Stationary Delta: 9.72197767280223e-06
#> Stationary step: 90
#> Stationary Delta: 9.53620734540726e-06
#> Stationary step: 86
#> Stationary Delta: 9.70249136101322e-06
#> Stationary step: 90
#> Stationary Delta: 9.95661423787166e-06
#> Stationary step: 88
#> Stationary Delta: 9.55863820927884e-06
#> Stationary step: 91
#> Stationary Delta: 9.47520460887182e-06
#> Stationary step: 89
#> Stationary Delta: 9.88100251343998e-06
#> Stationary step: 89
#> Stationary Delta: 9.76422905332985e-06
#> Stationary step: 89
#> Stationary Delta: 9.37272749479272e-06
#> Stationary step: 89
#> Stationary Delta: 9.5224237847622e-06
#> Stationary step: 90
#> Stationary Delta: 9.99484900293085e-06
#> Stationary step: 89
#> Stationary Delta: 9.3749242428783e-06
#> Stationary step: 89
#> Stationary Delta: 9.42586623597153e-06
#> Stationary step: 88
#> Stationary Delta: 9.5365455224938e-06
#> Stationary step: 86
#> Stationary Delta: 9.56850150246095e-06
#> Stationary step: 84
#> Stationary Delta: 9.79224772991888e-06
#> Stationary step: 89
#> Stationary Delta: 9.89177501847275e-06
#> Stationary step: 86
#> Stationary Delta: 9.76520948818012e-06
#> Stationary step: 91
#> Stationary Delta: 9.45760014989277e-06
#> Stationary step: 89
#> Stationary Delta: 9.60757633267641e-06
#> Stationary step: 91
#> Stationary Delta: 9.98764231167238e-06
#> Stationary step: 89
#> Stationary Delta: 9.95536252067023e-06
#> Stationary step: 85
#> Stationary Delta: 9.85143202897283e-06
#> Stationary step: 88
#> Stationary Delta: 9.4174963544671e-06
#> Stationary step: 86
#> Stationary Delta: 9.53793047550419e-06
#> Stationary step: 91
#> Stationary Delta: 9.35704083722575e-06
#> Stationary step: 87
#> Stationary Delta: 9.8985908730458e-06
#> Stationary step: 87
#> Stationary Delta: 9.45100592825839e-06
#> Stationary step: 91
#> Stationary Delta: 9.39959685963297e-06
#> Stationary step: 93
#> Stationary Delta: 9.79307060810198e-06
#> Stationary step: 91
#> Stationary Delta: 9.65972627058188e-06
#> Stationary step: 87
#> Stationary Delta: 9.63404693631189e-06
#> Stationary step: 91
#> Stationary Delta: 9.48013195045186e-06
#> Stationary step: 87
#> Stationary Delta: 9.82674201316009e-06
#> Stationary step: 89
#> Stationary Delta: 9.93362511247929e-06
#> Stationary step: 91
#> Stationary Delta: 9.33412922355165e-06
#> Stationary step: 84
#> Stationary Delta: 9.42648187690356e-06
#> Stationary step: 87
#> Stationary Delta: 9.91969963518938e-06
#> Stationary step: 89
#> Stationary Delta: 9.98102107970012e-06
#> Stationary step: 89
#> Stationary Delta: 9.43107742327122e-06
#> Stationary step: 91
#> Stationary Delta: 9.64951372893932e-06
#> Stationary step: 87
#> Stationary Delta: 9.78158456381053e-06
#> Stationary step: 94
#> Stationary Delta: 9.5562083637025e-06
#> Stationary step: 90
#> Stationary Delta: 9.40249230227673e-06
#> Stationary step: 90
#> Stationary Delta: 9.33869318625874e-06
#> Stationary step: 88
#> Stationary Delta: 9.82195337632729e-06
#> Stationary step: 87
#> Stationary Delta: 9.7601138172698e-06
#> Stationary step: 84
#> Stationary Delta: 9.99769667928174e-06
#> Stationary step: 86
#> Stationary Delta: 9.97321124202606e-06
#> Stationary step: 86
#> Stationary Delta: 9.76271994687448e-06
#> Stationary step: 87
#> Stationary Delta: 9.63705188857828e-06
#> Stationary step: 87
#> Stationary Delta: 9.86170076406036e-06
#> Stationary step: 87
#> Stationary Delta: 9.53527182714243e-06
#> Stationary step: 87
#> Stationary Delta: 9.660568737106e-06
#> Stationary step: 87
#> Stationary Delta: 9.5430940121557e-06
#> Stationary step: 87
#> Stationary Delta: 9.92916346424389e-06
#> Stationary step: 88
#> Stationary Delta: 9.55417917463347e-06
#> Stationary step: 88
#> Stationary Delta: 9.87304346653506e-06
#> Stationary step: 87
#> Stationary Delta: 9.78618236411141e-06
#> Stationary step: 90
#> Stationary Delta: 9.72969800527721e-06
#> Stationary step: 91
#> Stationary Delta: 9.89523680036e-06
#> Stationary step: 87
#> Stationary Delta: 9.68921949243541e-06
#> Stationary step: 90
#> Stationary Delta: 9.35689892774348e-06
#> Stationary step: 85
#> Stationary Delta: 9.85809830082291e-06
#> Stationary step: 88
#> Stationary Delta: 9.91840108287147e-06
#> Stationary step: 84
#> Stationary Delta: 9.74081702056256e-06
#> Stationary step: 87
#> Stationary Delta: 9.616069368221e-06
#> Stationary step: 87
#> Stationary Delta: 9.56777329162263e-06
#> Stationary step: 87
#> Stationary Delta: 9.47053544570821e-06
#> Stationary step: 87
#> Stationary Delta: 9.49103734903687e-06
#> Stationary step: 87
#> Stationary Delta: 9.85824250363509e-06
#> Stationary step: 90
#> Stationary Delta: 9.74194795467259e-06
#> Stationary step: 85
#> Stationary Delta: 9.86450651623971e-06
#> Stationary step: 87
#> Stationary Delta: 9.51856142105167e-06
#> Stationary step: 90
#> Stationary Delta: 9.88354091278864e-06
#> Stationary step: 84
#> Stationary Delta: 9.82888729793307e-06
#> Stationary step: 89
#> Stationary Delta: 9.57321484168838e-06
#> Stationary step: 90
#> Stationary Delta: 9.58718003011118e-06
#> 100
#> cells passed 0.001 threshold: 13.06%
#> cells passed 0.01 threshold: 16.68%
#> cells passed 0.05 threshold: 25.93%
#> your emprical P value threshold: 0.05
#> what propertion of enriched cells over all cells: 25.93%
#> what propertion of seed cells that are enriched cells: 97.81%
#> fold of true cell over seed: 5.16