get_sigcell_simple

a function to determine statistically trait-enriched cell by permutation test.

get_sigcell_simple(
  knn_sparse_mat,
  seed_idx,
  topseed_npscore,
  permutation_times = 1000,
  true_cell_significance = 0.05,
  rda_output = FALSE,
  out_rda = tempfile(fileext = "true_cell_df.rda"),
  mycores = 1,
  rw_gamma = 0.05
)

Arguments

knn_sparse_mat

a sparse matrix used for network propagation, which indicates the adjacent matrix (m x m, where m is the cell number) of cell-to-cell network (M-kNN graph)

seed_idx

a logical vector indicating seed cells (TRUE) and non-seed cells (FALSE) with length of m, where m is the cell number. The length and position are corresponding to knn_sparse_mat

topseed_npscore

a numeric vector of real network propagation score

permutation_times

an integer describe times of permutation test for each cell

true_cell_significance

a numeric value between 0-1 indicating the significant threshold used to determine statistically trait-enriched cell

rda_output

if output details of each permutation as a rda

out_rda

if rda_output=T, an rda will be outputed with path and name specified

mycores

how many cores used for permutation test

rw_gamma

gamma from randomWalk_sparse function. Keep the same with what you used in the real setting

Value

A dataframe with two columns, the first one is seed index (the same with input), the second one is significance from permutation test. In addition, an R data is generated, which contains this dataframe and another dataframe depicting the network propagation score for each cell at each permuation.

Examples

knn_sparse_mat <- example_results$mutualknn30
mono_mat <- example_results$mono_mat
ture_cell_df <- get_sigcell_simple(knn_sparse_mat=knn_sparse_mat,
                                   seed_idx=mono_mat$seed_idx,
                                   topseed_npscore=mono_mat$np_score,
                                   permutation_times=100)
#> Get started!
#> Stationary step: 82
#> Stationary Delta: 9.96599441731866e-06
#> Stationary step: 87
#> Stationary Delta: 9.97610019925462e-06
#> Stationary step: 92
#> Stationary Delta: 9.94381212155877e-06
#> Stationary step: 90
#> Stationary Delta: 9.81598349177524e-06
#> Stationary step: 91
#> Stationary Delta: 9.56374738144406e-06
#> Stationary step: 89
#> Stationary Delta: 9.97872814528289e-06
#> Stationary step: 89
#> Stationary Delta: 9.4261350042988e-06
#> Stationary step: 91
#> Stationary Delta: 9.45302522154671e-06
#> Stationary step: 87
#> Stationary Delta: 9.86866424171455e-06
#> Stationary step: 88
#> Stationary Delta: 9.51075835907328e-06
#> Stationary step: 88
#> Stationary Delta: 9.58630735253963e-06
#> Stationary step: 90
#> Stationary Delta: 9.41506271690978e-06
#> Stationary step: 88
#> Stationary Delta: 9.66731847488808e-06
#> Stationary step: 87
#> Stationary Delta: 9.99069539431481e-06
#> Stationary step: 88
#> Stationary Delta: 9.82086691598869e-06
#> Stationary step: 88
#> Stationary Delta: 9.77050776969199e-06
#> Stationary step: 84
#> Stationary Delta: 9.86975263781689e-06
#> Stationary step: 90
#> Stationary Delta: 9.80286499393411e-06
#> Stationary step: 87
#> Stationary Delta: 9.7856114668152e-06
#> Stationary step: 92
#> Stationary Delta: 9.81009140061978e-06
#> Stationary step: 95
#> Stationary Delta: 9.41233615925655e-06
#> Stationary step: 89
#> Stationary Delta: 9.77940791627356e-06
#> Stationary step: 88
#> Stationary Delta: 9.56288462592116e-06
#> Stationary step: 88
#> Stationary Delta: 9.4132365922131e-06
#> Stationary step: 83
#> Stationary Delta: 9.58043649272293e-06
#> Stationary step: 90
#> Stationary Delta: 9.36252017496248e-06
#> Stationary step: 87
#> Stationary Delta: 9.40570131157015e-06
#> Stationary step: 89
#> Stationary Delta: 9.85850750116877e-06
#> Stationary step: 91
#> Stationary Delta: 9.50591425709634e-06
#> Stationary step: 88
#> Stationary Delta: 9.97103852758293e-06
#> Stationary step: 90
#> Stationary Delta: 9.73570281557032e-06
#> Stationary step: 87
#> Stationary Delta: 9.50750574475867e-06
#> Stationary step: 89
#> Stationary Delta: 9.35116731970451e-06
#> Stationary step: 87
#> Stationary Delta: 9.71244996114478e-06
#> Stationary step: 88
#> Stationary Delta: 9.97453638330788e-06
#> Stationary step: 87
#> Stationary Delta: 9.6288201367149e-06
#> Stationary step: 87
#> Stationary Delta: 9.54381510943071e-06
#> Stationary step: 92
#> Stationary Delta: 9.62185642381715e-06
#> Stationary step: 91
#> Stationary Delta: 9.77416413950409e-06
#> Stationary step: 92
#> Stationary Delta: 9.38999109708971e-06
#> Stationary step: 93
#> Stationary Delta: 9.65060049499947e-06
#> Stationary step: 86
#> Stationary Delta: 9.38798101836359e-06
#> Stationary step: 92
#> Stationary Delta: 9.67116053557575e-06
#> Stationary step: 87
#> Stationary Delta: 9.99699230824215e-06
#> Stationary step: 87
#> Stationary Delta: 9.99406423787781e-06
#> Stationary step: 89
#> Stationary Delta: 9.66890858662201e-06
#> Stationary step: 90
#> Stationary Delta: 9.37643835131849e-06
#> Stationary step: 88
#> Stationary Delta: 9.66062624826588e-06
#> Stationary step: 90
#> Stationary Delta: 9.98417445396324e-06
#> Stationary step: 92
#> Stationary Delta: 9.59401435453094e-06
#> Stationary step: 89
#> Stationary Delta: 9.39369737348401e-06
#> Stationary step: 91
#> Stationary Delta: 9.98306979203433e-06
#> Stationary step: 89
#> Stationary Delta: 9.49093601409007e-06
#> Stationary step: 91
#> Stationary Delta: 9.69041235424605e-06
#> Stationary step: 86
#> Stationary Delta: 9.63428654808726e-06
#> Stationary step: 91
#> Stationary Delta: 9.43692444200848e-06
#> Stationary step: 89
#> Stationary Delta: 9.98555891464593e-06
#> Stationary step: 88
#> Stationary Delta: 9.44183026112021e-06
#> Stationary step: 86
#> Stationary Delta: 9.7339827192774e-06
#> Stationary step: 88
#> Stationary Delta: 9.63122545961661e-06
#> Stationary step: 86
#> Stationary Delta: 9.80536171477607e-06
#> Stationary step: 82
#> Stationary Delta: 9.99568114681697e-06
#> Stationary step: 88
#> Stationary Delta: 9.78309438608415e-06
#> Stationary step: 91
#> Stationary Delta: 9.35314891170853e-06
#> Stationary step: 90
#> Stationary Delta: 9.47382627701067e-06
#> Stationary step: 88
#> Stationary Delta: 9.732903762743e-06
#> Stationary step: 88
#> Stationary Delta: 9.7214691867913e-06
#> Stationary step: 89
#> Stationary Delta: 9.83146326341828e-06
#> Stationary step: 89
#> Stationary Delta: 9.86206798187301e-06
#> Stationary step: 93
#> Stationary Delta: 9.8188346378905e-06
#> Stationary step: 91
#> Stationary Delta: 9.75861459509286e-06
#> Stationary step: 88
#> Stationary Delta: 9.96702745797843e-06
#> Stationary step: 87
#> Stationary Delta: 9.54474974413256e-06
#> Stationary step: 87
#> Stationary Delta: 9.5956262513274e-06
#> Stationary step: 84
#> Stationary Delta: 9.62988670747378e-06
#> Stationary step: 89
#> Stationary Delta: 9.6161779638668e-06
#> Stationary step: 88
#> Stationary Delta: 9.97417021032215e-06
#> Stationary step: 86
#> Stationary Delta: 9.40277801647274e-06
#> Stationary step: 85
#> Stationary Delta: 9.72443887797568e-06
#> Stationary step: 86
#> Stationary Delta: 9.94211936152924e-06
#> Stationary step: 87
#> Stationary Delta: 9.81608091341652e-06
#> Stationary step: 91
#> Stationary Delta: 9.68827181701831e-06
#> Stationary step: 89
#> Stationary Delta: 9.34424228606977e-06
#> Stationary step: 86
#> Stationary Delta: 9.55066052723266e-06
#> Stationary step: 88
#> Stationary Delta: 9.68184742568465e-06
#> Stationary step: 92
#> Stationary Delta: 9.52169081618012e-06
#> Stationary step: 90
#> Stationary Delta: 9.43074221287313e-06
#> Stationary step: 92
#> Stationary Delta: 9.65853064854873e-06
#> Stationary step: 88
#> Stationary Delta: 9.37058997142417e-06
#> Stationary step: 87
#> Stationary Delta: 9.5851056425353e-06
#> Stationary step: 90
#> Stationary Delta: 9.34136754334235e-06
#> Stationary step: 91
#> Stationary Delta: 9.39040228690588e-06
#> Stationary step: 88
#> Stationary Delta: 9.7448258247676e-06
#> Stationary step: 88
#> Stationary Delta: 9.58484417915987e-06
#> Stationary step: 91
#> Stationary Delta: 9.94949766691268e-06
#> Stationary step: 87
#> Stationary Delta: 9.95497753460633e-06
#> Stationary step: 88
#> Stationary Delta: 9.95701106322735e-06
#> Stationary step: 90
#> Stationary Delta: 9.48301308846629e-06
#> Stationary step: 87
#> Stationary Delta: 9.69921410590509e-06
#> Stationary step: 92
#> Stationary Delta: 9.94909183258284e-06
#> 100
#> cells passed 0.001 threshold: 13.57%
#> cells passed 0.01 threshold: 16.88%
#> cells passed 0.05 threshold: 25.86%
#> your emprical P value threshold: 0.05
#> what propertion of enriched cells over all cells: 25.86%
#> what propertion of seed cells that are enriched cells: 97.81%
#> fold of true cell over seed: 5.15