get_sigcell_simple

a function to determine statistically trait-enriched cell by permutation test.

get_sigcell_simple(
  knn_sparse_mat,
  seed_idx,
  topseed_npscore,
  permutation_times = 1000,
  true_cell_significance = 0.05,
  rda_output = FALSE,
  out_rda = tempfile(fileext = "true_cell_df.rda"),
  mycores = 1,
  rw_gamma = 0.05
)

Arguments

knn_sparse_mat

a sparse matrix used for network propagation, which indicates the adjacent matrix (m x m, where m is the cell number) of cell-to-cell network (M-kNN graph)

seed_idx

a logical vector indicating seed cells (TRUE) and non-seed cells (FALSE) with length of m, where m is the cell number. The length and position are corresponding to knn_sparse_mat

topseed_npscore

a numeric vector of real network propagation score

permutation_times

an integer describe times of permutation test for each cell

true_cell_significance

a numeric value between 0-1 indicating the significant threshold used to determine statistically trait-enriched cell

rda_output

if output details of each permutation as a rda

out_rda

if rda_output=T, an rda will be outputed with path and name specified

mycores

how many cores used for permutation test

rw_gamma

gamma from randomWalk_sparse function. Keep the same with what you used in the real setting

Value

A dataframe with two columns, the first one is seed index (the same with input), the second one is significance from permutation test. In addition, an R data is generated, which contains this dataframe and another dataframe depicting the network propagation score for each cell at each permuation.

Examples

knn_sparse_mat <- example_results$mutualknn30
mono_mat <- example_results$mono_mat
ture_cell_df <- get_sigcell_simple(knn_sparse_mat=knn_sparse_mat,
                                   seed_idx=mono_mat$seed_idx,
                                   topseed_npscore=mono_mat$np_score,
                                   permutation_times=100)
#> Get started!
#> Stationary step: 89
#> Stationary Delta: 9.30239589115659e-06
#> Stationary step: 85
#> Stationary Delta: 9.68519319127391e-06
#> Stationary step: 89
#> Stationary Delta: 9.56351742224396e-06
#> Stationary step: 87
#> Stationary Delta: 9.58155613372655e-06
#> Stationary step: 89
#> Stationary Delta: 9.70265766381462e-06
#> Stationary step: 86
#> Stationary Delta: 9.8518780646317e-06
#> Stationary step: 87
#> Stationary Delta: 9.90881367333682e-06
#> Stationary step: 87
#> Stationary Delta: 9.88374539443406e-06
#> Stationary step: 84
#> Stationary Delta: 9.98085217518353e-06
#> Stationary step: 82
#> Stationary Delta: 9.97746203953097e-06
#> Stationary step: 88
#> Stationary Delta: 9.76463839496834e-06
#> Stationary step: 90
#> Stationary Delta: 9.96057996301818e-06
#> Stationary step: 92
#> Stationary Delta: 9.48496621186767e-06
#> Stationary step: 88
#> Stationary Delta: 9.5573920161348e-06
#> Stationary step: 83
#> Stationary Delta: 9.50150557946251e-06
#> Stationary step: 91
#> Stationary Delta: 9.73495126886948e-06
#> Stationary step: 90
#> Stationary Delta: 9.78910140106438e-06
#> Stationary step: 88
#> Stationary Delta: 9.85736132538988e-06
#> Stationary step: 87
#> Stationary Delta: 9.90670640145493e-06
#> Stationary step: 91
#> Stationary Delta: 9.86280787113684e-06
#> Stationary step: 89
#> Stationary Delta: 9.49866194818117e-06
#> Stationary step: 91
#> Stationary Delta: 9.40839439123887e-06
#> Stationary step: 88
#> Stationary Delta: 9.88198204439096e-06
#> Stationary step: 82
#> Stationary Delta: 9.83113150269936e-06
#> Stationary step: 88
#> Stationary Delta: 9.73026849236166e-06
#> Stationary step: 89
#> Stationary Delta: 9.91188994881585e-06
#> Stationary step: 85
#> Stationary Delta: 9.95357997112374e-06
#> Stationary step: 87
#> Stationary Delta: 9.43834990988227e-06
#> Stationary step: 86
#> Stationary Delta: 9.84158347959945e-06
#> Stationary step: 86
#> Stationary Delta: 9.5912025865144e-06
#> Stationary step: 88
#> Stationary Delta: 9.34569414938663e-06
#> Stationary step: 88
#> Stationary Delta: 9.81845725681555e-06
#> Stationary step: 87
#> Stationary Delta: 9.6885560216048e-06
#> Stationary step: 86
#> Stationary Delta: 9.8017926262053e-06
#> Stationary step: 88
#> Stationary Delta: 9.38025723451302e-06
#> Stationary step: 86
#> Stationary Delta: 9.39475720663452e-06
#> Stationary step: 89
#> Stationary Delta: 9.37057498271728e-06
#> Stationary step: 90
#> Stationary Delta: 9.33676107088898e-06
#> Stationary step: 91
#> Stationary Delta: 9.63594995319123e-06
#> Stationary step: 87
#> Stationary Delta: 9.4647123315538e-06
#> Stationary step: 92
#> Stationary Delta: 9.33400319534483e-06
#> Stationary step: 84
#> Stationary Delta: 9.67420799181299e-06
#> Stationary step: 84
#> Stationary Delta: 9.8679289095856e-06
#> Stationary step: 90
#> Stationary Delta: 9.9408512788638e-06
#> Stationary step: 88
#> Stationary Delta: 9.45878971574318e-06
#> Stationary step: 88
#> Stationary Delta: 9.36739920734273e-06
#> Stationary step: 88
#> Stationary Delta: 9.67402504087586e-06
#> Stationary step: 90
#> Stationary Delta: 9.5079256449002e-06
#> Stationary step: 90
#> Stationary Delta: 9.40011520376079e-06
#> Stationary step: 86
#> Stationary Delta: 9.32245077610952e-06
#> Stationary step: 87
#> Stationary Delta: 9.81963069433631e-06
#> Stationary step: 87
#> Stationary Delta: 9.72368576889837e-06
#> Stationary step: 89
#> Stationary Delta: 9.62112243936777e-06
#> Stationary step: 88
#> Stationary Delta: 9.6333055751761e-06
#> Stationary step: 91
#> Stationary Delta: 9.59454426984843e-06
#> Stationary step: 87
#> Stationary Delta: 9.62460989274308e-06
#> Stationary step: 90
#> Stationary Delta: 9.61438981449068e-06
#> Stationary step: 90
#> Stationary Delta: 9.82024992166661e-06
#> Stationary step: 89
#> Stationary Delta: 9.72073499069065e-06
#> Stationary step: 89
#> Stationary Delta: 9.92314667177546e-06
#> Stationary step: 90
#> Stationary Delta: 9.35199917522054e-06
#> Stationary step: 88
#> Stationary Delta: 9.83248321149289e-06
#> Stationary step: 89
#> Stationary Delta: 9.42659531111244e-06
#> Stationary step: 88
#> Stationary Delta: 9.59498743605902e-06
#> Stationary step: 89
#> Stationary Delta: 9.65981719095993e-06
#> Stationary step: 84
#> Stationary Delta: 9.82124346909764e-06
#> Stationary step: 88
#> Stationary Delta: 9.46736629913874e-06
#> Stationary step: 92
#> Stationary Delta: 9.59415205077747e-06
#> Stationary step: 90
#> Stationary Delta: 9.96916930257029e-06
#> Stationary step: 91
#> Stationary Delta: 9.3401004693946e-06
#> Stationary step: 90
#> Stationary Delta: 9.61882303910808e-06
#> Stationary step: 88
#> Stationary Delta: 9.54852221060987e-06
#> Stationary step: 88
#> Stationary Delta: 9.76890941051403e-06
#> Stationary step: 88
#> Stationary Delta: 9.88162268920608e-06
#> Stationary step: 85
#> Stationary Delta: 9.87887920127876e-06
#> Stationary step: 88
#> Stationary Delta: 9.92435877004026e-06
#> Stationary step: 88
#> Stationary Delta: 9.61724515903474e-06
#> Stationary step: 89
#> Stationary Delta: 9.47867303638152e-06
#> Stationary step: 86
#> Stationary Delta: 9.90881996345403e-06
#> Stationary step: 93
#> Stationary Delta: 9.92029691539695e-06
#> Stationary step: 88
#> Stationary Delta: 9.3555940827902e-06
#> Stationary step: 83
#> Stationary Delta: 9.90507890299171e-06
#> Stationary step: 94
#> Stationary Delta: 9.54359080655413e-06
#> Stationary step: 92
#> Stationary Delta: 9.55911034672639e-06
#> Stationary step: 89
#> Stationary Delta: 9.74223104576294e-06
#> Stationary step: 86
#> Stationary Delta: 9.54481178132908e-06
#> Stationary step: 92
#> Stationary Delta: 9.92922328838542e-06
#> Stationary step: 92
#> Stationary Delta: 9.95414608527375e-06
#> Stationary step: 86
#> Stationary Delta: 9.43523329248537e-06
#> Stationary step: 89
#> Stationary Delta: 9.424060223754e-06
#> Stationary step: 86
#> Stationary Delta: 9.79742209102523e-06
#> Stationary step: 89
#> Stationary Delta: 9.5192522992416e-06
#> Stationary step: 93
#> Stationary Delta: 9.8760989453492e-06
#> Stationary step: 89
#> Stationary Delta: 9.464359452382e-06
#> Stationary step: 87
#> Stationary Delta: 9.69269087318621e-06
#> Stationary step: 90
#> Stationary Delta: 9.34136220169203e-06
#> Stationary step: 94
#> Stationary Delta: 9.58721647495462e-06
#> Stationary step: 87
#> Stationary Delta: 9.76112831018199e-06
#> Stationary step: 90
#> Stationary Delta: 9.79038966589943e-06
#> Stationary step: 92
#> Stationary Delta: 9.53128417014644e-06
#> 100
#> cells passed 0.001 threshold: 13.35%
#> cells passed 0.01 threshold: 16.61%
#> cells passed 0.05 threshold: 25.91%
#> your emprical P value threshold: 0.05
#> what propertion of enriched cells over all cells: 25.91%
#> what propertion of seed cells that are enriched cells: 97.37%
#> fold of true cell over seed: 5.16