get_sigcell_simple

a function to determine statistically trait-enriched cell by permutation test.

get_sigcell_simple(
  knn_sparse_mat,
  seed_idx,
  topseed_npscore,
  permutation_times = 1000,
  true_cell_significance = 0.05,
  rda_output = FALSE,
  out_rda = tempfile(fileext = "true_cell_df.rda"),
  mycores = 1,
  rw_gamma = 0.05
)

Arguments

knn_sparse_mat

a sparse matrix used for network propagation, which indicates the adjacent matrix (m x m, where m is the cell number) of cell-to-cell network (M-kNN graph)

seed_idx

a logical vector indicating seed cells (TRUE) and non-seed cells (FALSE) with length of m, where m is the cell number. The length and position are corresponding to knn_sparse_mat

topseed_npscore

a numeric vector of real network propagation score

permutation_times

an integer describe times of permutation test for each cell

true_cell_significance

a numeric value between 0-1 indicating the significant threshold used to determine statistically trait-enriched cell

rda_output

if output details of each permutation as a rda

out_rda

if rda_output=T, an rda will be outputed with path and name specified

mycores

how many cores used for permutation test

rw_gamma

gamma from randomWalk_sparse function. Keep the same with what you used in the real setting

Value

A dataframe with two columns, the first one is seed index (the same with input), the second one is significance from permutation test. In addition, an R data is generated, which contains this dataframe and another dataframe depicting the network propagation score for each cell at each permuation.

Examples

knn_sparse_mat <- example_results$mutualknn30
mono_mat <- example_results$mono_mat
ture_cell_df <- get_sigcell_simple(knn_sparse_mat=knn_sparse_mat,
                                   seed_idx=mono_mat$seed_idx,
                                   topseed_npscore=mono_mat$np_score,
                                   permutation_times=100)
#> Get started!
#> Stationary step: 92
#> Stationary Delta: 9.44291400775169e-06
#> Stationary step: 90
#> Stationary Delta: 9.84232720690703e-06
#> Stationary step: 89
#> Stationary Delta: 9.71948389567192e-06
#> Stationary step: 86
#> Stationary Delta: 9.36286797726153e-06
#> Stationary step: 86
#> Stationary Delta: 9.8321227104769e-06
#> Stationary step: 90
#> Stationary Delta: 9.80222003502446e-06
#> Stationary step: 91
#> Stationary Delta: 9.5765816005416e-06
#> Stationary step: 92
#> Stationary Delta: 9.84458470547632e-06
#> Stationary step: 80
#> Stationary Delta: 9.945958213361e-06
#> Stationary step: 88
#> Stationary Delta: 9.43994067569369e-06
#> Stationary step: 83
#> Stationary Delta: 9.6074775067712e-06
#> Stationary step: 91
#> Stationary Delta: 9.8454879233651e-06
#> Stationary step: 90
#> Stationary Delta: 9.40345376360838e-06
#> Stationary step: 86
#> Stationary Delta: 9.4789920777993e-06
#> Stationary step: 88
#> Stationary Delta: 9.34246599802123e-06
#> Stationary step: 87
#> Stationary Delta: 9.44590745470113e-06
#> Stationary step: 91
#> Stationary Delta: 9.60350479256745e-06
#> Stationary step: 89
#> Stationary Delta: 9.47874166613584e-06
#> Stationary step: 92
#> Stationary Delta: 9.75302833936507e-06
#> Stationary step: 89
#> Stationary Delta: 9.3547761667557e-06
#> Stationary step: 87
#> Stationary Delta: 9.83057436125972e-06
#> Stationary step: 93
#> Stationary Delta: 9.78101338980488e-06
#> Stationary step: 88
#> Stationary Delta: 9.69030093600586e-06
#> Stationary step: 88
#> Stationary Delta: 9.43951746912215e-06
#> Stationary step: 90
#> Stationary Delta: 9.36271507816355e-06
#> Stationary step: 89
#> Stationary Delta: 9.47864695367499e-06
#> Stationary step: 83
#> Stationary Delta: 9.92376244725395e-06
#> Stationary step: 83
#> Stationary Delta: 9.86125920539013e-06
#> Stationary step: 85
#> Stationary Delta: 9.76616400151029e-06
#> Stationary step: 84
#> Stationary Delta: 9.92658471667667e-06
#> Stationary step: 91
#> Stationary Delta: 9.8706278435174e-06
#> Stationary step: 86
#> Stationary Delta: 9.61745533169791e-06
#> Stationary step: 88
#> Stationary Delta: 9.87231123028373e-06
#> Stationary step: 87
#> Stationary Delta: 9.34413612539221e-06
#> Stationary step: 91
#> Stationary Delta: 9.90718240911891e-06
#> Stationary step: 83
#> Stationary Delta: 9.97316238808776e-06
#> Stationary step: 88
#> Stationary Delta: 9.6151325581266e-06
#> Stationary step: 89
#> Stationary Delta: 9.77671516293678e-06
#> Stationary step: 88
#> Stationary Delta: 9.87223387957137e-06
#> Stationary step: 90
#> Stationary Delta: 9.9969499283527e-06
#> Stationary step: 92
#> Stationary Delta: 9.70935326986499e-06
#> Stationary step: 91
#> Stationary Delta: 9.68436504160518e-06
#> Stationary step: 86
#> Stationary Delta: 9.69759202007317e-06
#> Stationary step: 88
#> Stationary Delta: 9.75601164446643e-06
#> Stationary step: 92
#> Stationary Delta: 9.76074329382663e-06
#> Stationary step: 85
#> Stationary Delta: 9.34677211737343e-06
#> Stationary step: 90
#> Stationary Delta: 9.77811344662487e-06
#> Stationary step: 84
#> Stationary Delta: 9.50110977160076e-06
#> Stationary step: 87
#> Stationary Delta: 9.82662029922421e-06
#> Stationary step: 92
#> Stationary Delta: 9.52290489915207e-06
#> Stationary step: 83
#> Stationary Delta: 9.95437178804727e-06
#> Stationary step: 88
#> Stationary Delta: 9.99736327654986e-06
#> Stationary step: 88
#> Stationary Delta: 9.6937359807229e-06
#> Stationary step: 91
#> Stationary Delta: 9.71551634238766e-06
#> Stationary step: 90
#> Stationary Delta: 9.3230655632062e-06
#> Stationary step: 90
#> Stationary Delta: 9.75797680984563e-06
#> Stationary step: 87
#> Stationary Delta: 9.41490539223432e-06
#> Stationary step: 90
#> Stationary Delta: 9.80920176113527e-06
#> Stationary step: 89
#> Stationary Delta: 9.47475024085007e-06
#> Stationary step: 86
#> Stationary Delta: 9.50317301356999e-06
#> Stationary step: 87
#> Stationary Delta: 9.82383631481486e-06
#> Stationary step: 87
#> Stationary Delta: 9.49143498121593e-06
#> Stationary step: 90
#> Stationary Delta: 9.96118256406015e-06
#> Stationary step: 87
#> Stationary Delta: 9.33773543356814e-06
#> Stationary step: 88
#> Stationary Delta: 9.33510621349609e-06
#> Stationary step: 91
#> Stationary Delta: 9.71980678549611e-06
#> Stationary step: 84
#> Stationary Delta: 9.82526885701388e-06
#> Stationary step: 90
#> Stationary Delta: 9.88422199958098e-06
#> Stationary step: 90
#> Stationary Delta: 9.60326127919443e-06
#> Stationary step: 89
#> Stationary Delta: 9.54512060530862e-06
#> Stationary step: 85
#> Stationary Delta: 9.93005331736963e-06
#> Stationary step: 90
#> Stationary Delta: 9.6882774702716e-06
#> Stationary step: 89
#> Stationary Delta: 9.64079301692584e-06
#> Stationary step: 88
#> Stationary Delta: 9.68117667756581e-06
#> Stationary step: 88
#> Stationary Delta: 9.62894720550248e-06
#> Stationary step: 87
#> Stationary Delta: 9.30431731549862e-06
#> Stationary step: 90
#> Stationary Delta: 9.6499164070674e-06
#> Stationary step: 86
#> Stationary Delta: 9.74468208637125e-06
#> Stationary step: 86
#> Stationary Delta: 9.65579336830121e-06
#> Stationary step: 89
#> Stationary Delta: 9.72796659364058e-06
#> Stationary step: 89
#> Stationary Delta: 9.49745748689348e-06
#> Stationary step: 87
#> Stationary Delta: 9.78505733757656e-06
#> Stationary step: 89
#> Stationary Delta: 9.50034191325014e-06
#> Stationary step: 90
#> Stationary Delta: 9.47919613761446e-06
#> Stationary step: 91
#> Stationary Delta: 9.59889939273002e-06
#> Stationary step: 90
#> Stationary Delta: 9.49761043871285e-06
#> Stationary step: 89
#> Stationary Delta: 9.66894091890652e-06
#> Stationary step: 90
#> Stationary Delta: 9.85359000215747e-06
#> Stationary step: 88
#> Stationary Delta: 9.75969364924358e-06
#> Stationary step: 92
#> Stationary Delta: 9.58130020168627e-06
#> Stationary step: 88
#> Stationary Delta: 9.63126303340922e-06
#> Stationary step: 91
#> Stationary Delta: 9.84536132027129e-06
#> Stationary step: 91
#> Stationary Delta: 9.70299214264215e-06
#> Stationary step: 87
#> Stationary Delta: 9.87017212272638e-06
#> Stationary step: 88
#> Stationary Delta: 9.57687314277151e-06
#> Stationary step: 88
#> Stationary Delta: 9.92503165081699e-06
#> Stationary step: 86
#> Stationary Delta: 9.97742155626981e-06
#> Stationary step: 91
#> Stationary Delta: 9.54727604401213e-06
#> Stationary step: 91
#> Stationary Delta: 9.68256400298492e-06
#> Stationary step: 85
#> Stationary Delta: 9.83586788645762e-06
#> 100
#> cells passed 0.001 threshold: 13%
#> cells passed 0.01 threshold: 16.35%
#> cells passed 0.05 threshold: 25.64%
#> your emprical P value threshold: 0.05
#> what propertion of enriched cells over all cells: 25.64%
#> what propertion of seed cells that are enriched cells: 96.49%
#> fold of true cell over seed: 5.11