get_sigcell_simple

a function to determine statistically trait-enriched cell by permutation test.

get_sigcell_simple(
  knn_sparse_mat,
  seed_idx,
  topseed_npscore,
  permutation_times = 1000,
  true_cell_significance = 0.05,
  rda_output = FALSE,
  out_rda = tempfile(fileext = "true_cell_df.rda"),
  mycores = 1,
  rw_gamma = 0.05
)

Arguments

knn_sparse_mat

a sparse matrix used for network propagation, which indicates the adjacent matrix (m x m, where m is the cell number) of cell-to-cell network (M-kNN graph)

seed_idx

a logical vector indicating seed cells (TRUE) and non-seed cells (FALSE) with length of m, where m is the cell number. The length and position are corresponding to knn_sparse_mat

topseed_npscore

a numeric vector of real network propagation score

permutation_times

an integer describe times of permutation test for each cell

true_cell_significance

a numeric value between 0-1 indicating the significant threshold used to determine statistically trait-enriched cell

rda_output

if output details of each permutation as a rda

out_rda

if rda_output=T, an rda will be outputed with path and name specified

mycores

how many cores used for permutation test

rw_gamma

gamma from randomWalk_sparse function. Keep the same with what you used in the real setting

Value

A dataframe with two columns, the first one is seed index (the same with input), the second one is significance from permutation test. In addition, an R data is generated, which contains this dataframe and another dataframe depicting the network propagation score for each cell at each permuation.

Examples

knn_sparse_mat <- example_results$mutualknn30
mono_mat <- example_results$mono_mat
ture_cell_df <- get_sigcell_simple(knn_sparse_mat=knn_sparse_mat,
                                   seed_idx=mono_mat$seed_idx,
                                   topseed_npscore=mono_mat$np_score,
                                   permutation_times=100)
#> Get started!
#> Stationary step: 84
#> Stationary Delta: 9.77761588159129e-06
#> Stationary step: 89
#> Stationary Delta: 9.63906523284983e-06
#> Stationary step: 86
#> Stationary Delta: 9.68698527817392e-06
#> Stationary step: 85
#> Stationary Delta: 9.67529257365375e-06
#> Stationary step: 87
#> Stationary Delta: 9.32841252874703e-06
#> Stationary step: 88
#> Stationary Delta: 9.47450576534526e-06
#> Stationary step: 86
#> Stationary Delta: 9.82946460733495e-06
#> Stationary step: 89
#> Stationary Delta: 9.37159244970487e-06
#> Stationary step: 89
#> Stationary Delta: 9.77756404520008e-06
#> Stationary step: 86
#> Stationary Delta: 9.89921566036749e-06
#> Stationary step: 91
#> Stationary Delta: 9.66965276939918e-06
#> Stationary step: 90
#> Stationary Delta: 9.65658634661866e-06
#> Stationary step: 84
#> Stationary Delta: 9.68487810137621e-06
#> Stationary step: 90
#> Stationary Delta: 9.55207021430631e-06
#> Stationary step: 90
#> Stationary Delta: 9.29155685824122e-06
#> Stationary step: 85
#> Stationary Delta: 9.95876328483811e-06
#> Stationary step: 83
#> Stationary Delta: 9.46537649292549e-06
#> Stationary step: 85
#> Stationary Delta: 9.96516736287317e-06
#> Stationary step: 91
#> Stationary Delta: 9.56965387683672e-06
#> Stationary step: 84
#> Stationary Delta: 9.73203098524743e-06
#> Stationary step: 90
#> Stationary Delta: 9.51965523840491e-06
#> Stationary step: 87
#> Stationary Delta: 9.81766922058919e-06
#> Stationary step: 89
#> Stationary Delta: 9.5544437015011e-06
#> Stationary step: 89
#> Stationary Delta: 9.92049149738473e-06
#> Stationary step: 91
#> Stationary Delta: 9.46600587640424e-06
#> Stationary step: 89
#> Stationary Delta: 9.97886354349275e-06
#> Stationary step: 91
#> Stationary Delta: 9.49576827980329e-06
#> Stationary step: 92
#> Stationary Delta: 9.99014365776971e-06
#> Stationary step: 91
#> Stationary Delta: 9.73073174504212e-06
#> Stationary step: 87
#> Stationary Delta: 9.34239246115753e-06
#> Stationary step: 90
#> Stationary Delta: 9.41332507010087e-06
#> Stationary step: 89
#> Stationary Delta: 9.75824501656054e-06
#> Stationary step: 91
#> Stationary Delta: 9.59699556755664e-06
#> Stationary step: 93
#> Stationary Delta: 9.42863847198547e-06
#> Stationary step: 87
#> Stationary Delta: 9.59493577789879e-06
#> Stationary step: 89
#> Stationary Delta: 9.45455195306409e-06
#> Stationary step: 89
#> Stationary Delta: 9.89384459812258e-06
#> Stationary step: 92
#> Stationary Delta: 9.36917626674872e-06
#> Stationary step: 87
#> Stationary Delta: 9.73118576210178e-06
#> Stationary step: 88
#> Stationary Delta: 9.32724962760251e-06
#> Stationary step: 89
#> Stationary Delta: 9.80154258989714e-06
#> Stationary step: 90
#> Stationary Delta: 9.93445820428564e-06
#> Stationary step: 85
#> Stationary Delta: 9.36112795949494e-06
#> Stationary step: 90
#> Stationary Delta: 9.47601578880422e-06
#> Stationary step: 89
#> Stationary Delta: 9.47593853254564e-06
#> Stationary step: 89
#> Stationary Delta: 9.62930775507153e-06
#> Stationary step: 89
#> Stationary Delta: 9.31054723190639e-06
#> Stationary step: 86
#> Stationary Delta: 9.68488043628885e-06
#> Stationary step: 87
#> Stationary Delta: 9.82998848846717e-06
#> Stationary step: 86
#> Stationary Delta: 9.99052630372922e-06
#> Stationary step: 84
#> Stationary Delta: 9.91543968197115e-06
#> Stationary step: 88
#> Stationary Delta: 9.56198492476455e-06
#> Stationary step: 90
#> Stationary Delta: 9.69661146163037e-06
#> Stationary step: 91
#> Stationary Delta: 9.88766639135018e-06
#> Stationary step: 84
#> Stationary Delta: 9.5005948958616e-06
#> Stationary step: 79
#> Stationary Delta: 9.38906983750668e-06
#> Stationary step: 87
#> Stationary Delta: 9.78618758596885e-06
#> Stationary step: 86
#> Stationary Delta: 9.96090292924428e-06
#> Stationary step: 87
#> Stationary Delta: 9.48039175771687e-06
#> Stationary step: 87
#> Stationary Delta: 9.74475992112825e-06
#> Stationary step: 88
#> Stationary Delta: 9.94756355879258e-06
#> Stationary step: 87
#> Stationary Delta: 9.80613219231583e-06
#> Stationary step: 89
#> Stationary Delta: 9.71607758622005e-06
#> Stationary step: 90
#> Stationary Delta: 9.75019809217741e-06
#> Stationary step: 87
#> Stationary Delta: 9.51107666058769e-06
#> Stationary step: 87
#> Stationary Delta: 9.61270114869111e-06
#> Stationary step: 88
#> Stationary Delta: 9.78241911509518e-06
#> Stationary step: 90
#> Stationary Delta: 9.58741579079369e-06
#> Stationary step: 87
#> Stationary Delta: 9.80390101276222e-06
#> Stationary step: 88
#> Stationary Delta: 9.8597040322044e-06
#> Stationary step: 92
#> Stationary Delta: 9.4933731191907e-06
#> Stationary step: 87
#> Stationary Delta: 9.93234899073489e-06
#> Stationary step: 89
#> Stationary Delta: 9.30154550403317e-06
#> Stationary step: 93
#> Stationary Delta: 9.35425145345124e-06
#> Stationary step: 92
#> Stationary Delta: 9.93360399043583e-06
#> Stationary step: 89
#> Stationary Delta: 9.33842794712384e-06
#> Stationary step: 86
#> Stationary Delta: 9.4715636196019e-06
#> Stationary step: 87
#> Stationary Delta: 9.58853785725158e-06
#> Stationary step: 89
#> Stationary Delta: 9.54869950224715e-06
#> Stationary step: 87
#> Stationary Delta: 9.8898443870061e-06
#> Stationary step: 89
#> Stationary Delta: 9.6753566178598e-06
#> Stationary step: 92
#> Stationary Delta: 9.4437617340121e-06
#> Stationary step: 89
#> Stationary Delta: 9.91754267173145e-06
#> Stationary step: 86
#> Stationary Delta: 9.97439359988226e-06
#> Stationary step: 91
#> Stationary Delta: 9.60747467790063e-06
#> Stationary step: 86
#> Stationary Delta: 9.65754113870463e-06
#> Stationary step: 88
#> Stationary Delta: 9.33092044020856e-06
#> Stationary step: 90
#> Stationary Delta: 9.90616544658809e-06
#> Stationary step: 88
#> Stationary Delta: 9.92733417626786e-06
#> Stationary step: 92
#> Stationary Delta: 9.57699338932701e-06
#> Stationary step: 89
#> Stationary Delta: 9.69687165378165e-06
#> Stationary step: 88
#> Stationary Delta: 9.54391377223593e-06
#> Stationary step: 93
#> Stationary Delta: 9.65771590341882e-06
#> Stationary step: 90
#> Stationary Delta: 9.72041578428636e-06
#> Stationary step: 93
#> Stationary Delta: 9.8739969163836e-06
#> Stationary step: 88
#> Stationary Delta: 9.62340894662113e-06
#> Stationary step: 89
#> Stationary Delta: 9.93366259101981e-06
#> Stationary step: 86
#> Stationary Delta: 9.64971344421774e-06
#> Stationary step: 87
#> Stationary Delta: 9.78202856899291e-06
#> Stationary step: 90
#> Stationary Delta: 9.93792401202983e-06
#> 100
#> cells passed 0.001 threshold: 13.66%
#> cells passed 0.01 threshold: 16.83%
#> cells passed 0.05 threshold: 26.13%
#> your emprical P value threshold: 0.05
#> what propertion of enriched cells over all cells: 26.13%
#> what propertion of seed cells that are enriched cells: 97.37%
#> fold of true cell over seed: 5.2